Undo respond_with removals.
This is causing two problems.

The first is that all URLs inside controllers, helpers and views are now needed to be prefixed with the spree routing helper. This is not necessary, as you can see from [this line][1] within Forem.

The second problem is that URLs *without* this prefix can cause the tests memory consumption to balloon and consume all the memory on the machine. This is why the Travis tests were being killed: https://travis-ci.org/#!/spree/spree/jobs/2969163

[1]: https://github.com/radar/forem/blob/2fcf21adf1d4e44e2311862b035c79f2660b4a76/app/controllers/forem/admin/forums_controller.rb#L18

This is a combination of 4 reversion commits.
Revert "Remove respond_with helper"

This reverts commit e1920f638b498964bd58fe79e3c269ad4d8f5a1d.

Revert "Remove respond_{to/with} calls from controllers that don't need them"

This reverts commit 4e0249dbaa8d31d682bc9133f53411fe57aab3b5.

Conflicts:

	core/app/controllers/spree/checkout_controller.rb
	core/app/controllers/spree/orders_controller.rb

Revert "Remove Spree::Core::Responder"

This reverts commit 992683015986e3a899f4ced70a55a5f8b0f93b6f.

Revert "Remove require for spree/core/responder"

This reverts commit dab1528d9d8e3b39f221a189691d4994e19b4cd8.